Course course service

+++ Backend/PreMatricula/src/main/java/com/example/preMatricula/interfaces/DisciplinaRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
  
 @Repository
-public interface DisciplinaRepository extends MongoRepository<Discipline, Long> {
+public interface DisciplinaRepository extends MongoRepository<Discipline, Integer> {
 
 }
